+++ client/src/hooks/helpers/useRealm.tsx
@@ -31,7 +31,7 @@ export function useRealm() {
   } = useDojo();
 
   const getQuestResources = () => {
-    const realmEntityId = useUIStore.getState().realmEntityId;
+    const realmEntityId = useUIStore((state) => state.realmEntityId);
     const realm = getComponentValue(Realm, getEntityIdFromKeys([BigInt(realmEntityId)]));
     const resourcesProduced = realm ? unpackResources(realm.resource_types_packed, realm.resource_types_count) : [];
     return getStartingResources(resourcesProduced);
